Add `_stbt.types` containing types with no dependencies
This is part of an effort to split up `core.py` into smaller chunks.

As a side-effect `Region.bounding_box` is now a public API.
